# Fix prompt tuning with left padding