--- a/src/peft/peft_model.py
+++ b/src/peft/peft_model.py
@@ -21,7 +21,7 @@ import os
 import warnings
 from contextlib import contextmanager
 from copy import deepcopy
-from typing import Any, Dict, List, Optional, Union, Tuple
+from typing import Any, Dict, List, Optional, Union
 
 import torch
 from accelerate import dispatch_model, infer_auto_device_map
@@ -59,14 +59,14 @@ from .utils import (
     _prepare_prompt_learning_config,
     _set_adapter,
     _set_trainable,
+    add_pad_tokens,
     get_peft_model_state_dict,
     id_tensor_storage,
     infer_device,
     load_peft_weights,
+    separate_pad_tokens,
     set_peft_model_state_dict,
     shift_tokens_right,
-    separate_pad_tokens,
-    add_pad_tokens,
 )
 
 
@@ -1036,12 +1036,17 @@ class PeftModelForCausalLM(PeftModel):
 
         batch_size = _get_batch_size(input_ids, inputs_embeds)
         if not peft_config.peft_type == PeftType.PREFIX_TUNING:
-            input_ids, inputs_embeds, attention_mask, labels, pad_els = separate_pad_tokens(input_ids, inputs_embeds, attention_mask, labels)
+            input_ids, inputs_embeds, attention_mask, labels, pad_els = separate_pad_tokens(
+                input_ids, inputs_embeds, attention_mask, labels
+            )
         if attention_mask is not None:
             # concat prompt attention mask. index into first element for device info since attentin_mask can be a list for prompt tuning
             prefix_attention_mask = torch.ones(batch_size, peft_config.num_virtual_tokens).to(attention_mask[0].device)
             if not peft_config.peft_type == PeftType.PREFIX_TUNING:
-                attention_mask = [torch.cat((prefix_attention_mask_single, mask)) for prefix_attention_mask_single, mask in zip(prefix_attention_mask, attention_mask)]
+                attention_mask = [
+                    torch.cat((prefix_attention_mask_single, mask))
+                    for prefix_attention_mask_single, mask in zip(prefix_attention_mask, attention_mask)
+                ]
             else:
                 attention_mask = torch.cat((prefix_attention_mask, attention_mask), dim=1)
 
@@ -1076,10 +1081,10 @@ class PeftModelForCausalLM(PeftModel):
                 labels = [torch.cat((prefix_label, label)) for prefix_label, label in zip(prefix_labels, labels)]
             prompts = self.get_prompt(batch_size=batch_size, task_ids=task_ids)
             prompts = prompts.to(inputs_embeds[0].dtype)
-            
+
             inputs_embeds = [torch.cat((prompt, input_embed)) for prompt, input_embed in zip(prompts, inputs_embeds)]
             input_embed_paddings = pad_els[1]
-            if not input_embed_paddings: 
+            if not input_embed_paddings:
                 # get padding token embeddings
                 padding_embeds = [self.word_embeddings(pad_id) for pad_id in pad_els[0]]
                 pad_els = (pad_els[0], padding_embeds, pad_els[2], pad_els[3])