--- a/src/peft/utils/__init__.py
+++ b/src/peft/utils/__init__.py
@@ -46,5 +46,7 @@ from .other import (
     get_auto_gptq_quant_linear,
     get_quantization_config,
     id_tensor_storage,
+    separate_pad_tokens,
+    add_pad_tokens,
 )
 from .save_and_load import get_peft_model_state_dict, set_peft_model_state_dict, load_peft_weights