--- a/src/peft/utils/other.py
+++ b/src/peft/utils/other.py
@@ -15,7 +15,7 @@
 import copy
 import inspect
 import warnings
-from typing import Optional, Tuple
+from typing import Optional, Tuple, List
 
 import accelerate
 import torch
@@ -476,6 +476,56 @@ def id_tensor_storage(tensor: torch.Tensor) -> Tuple[torch.device, int, int]:
         unique_id = storage_ptr(tensor)
 
     return tensor.device, unique_id, storage_size(tensor)
+
+def separate_pad_tokens(input_ids: torch.Tensor, attention_mask: torch.Tensor, labels: torch.Tensor):
+    """
+    Separates out the tokens corresponding to padding in input_ids, attention_mask and labels
+    Returns:
+    - input_ids: input_ids with padding tokens removed
+    - attention_mask: attention_mask with padding tokens removed
+    - labels: labels with padding tokens removed
+    - pad_els: a tuple (pad_token_input_ids, pad_token_attention_mask, pad_token_labels) containing the padding token elements
+    """
+    if not input_ids.ndim == 2:
+        raise ValueError("input_ids must be a 2D tensor")
+    # find the first index of a non-padding token
+    if attention_mask is not None:
+        first_non_padding_tok_ind = torch.argmax(attention_mask, dim=1)
+    else:
+        first_non_padding_tok_ind = torch.zeros(input_ids.shape[0], dtype=torch.long, device=input_ids.device)
+    input_paddings = [input_ids[i][:first_ind] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+    # get the same slices for attention_mask and labels. Note that all the padding vectors are of different shapes 
+    # so we can't stack them into a tensor
+    attention_mask_paddings, labels_paddings = None, None
+    if attention_mask is not None:
+        attention_mask_paddings = [attention_mask[i][:first_ind] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+        attention_mask = [attention_mask[i][first_ind:] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+    
+    if labels is not None:
+        labels_paddings = [labels[i][:first_ind] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+        labels = [labels[i][first_ind:] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+
+    pad_els = (input_paddings, attention_mask_paddings, labels_paddings)
+    input_ids = [input_ids[i][first_ind:] for i, first_ind in enumerate(first_non_padding_tok_ind)]
+    return input_ids, attention_mask, labels, pad_els
+
+def add_pad_tokens(input_embeds: List[torch.Tensor], attention_mask: List[torch.Tensor], labels: List[torch.Tensor], pad_els: Tuple[List[torch.Tensor], List[torch.Tensor], List[torch.Tensor]]):
+    """
+    Adds the padding token elements back to input_ids, attention_mask and labels
+    Returns:
+    - input_ids: input_ids with padding tokens removed
+    - attention_mask: attention_mask with padding tokens removed
+    - labels: labels with padding tokens removed
+    - pad_els: a tuple (pad_token_input_ids, pad_token_attention_mask, pad_token_labels) containing the padding tokens
+    """
+    pad_embeds, pad_mask, pad_labels = pad_els
+    final_input_embeds = torch.stack([torch.cat((pad_embed, input_embed)) for input_embed, pad_embed in zip(input_embeds, pad_embeds)])
+    final_attention_mask, final_labels = None, None
+    if attention_mask is not None:
+        final_attention_mask = torch.stack([torch.cat((pad_mask, attention_mask)) for attention_mask, pad_mask in zip(attention_mask, pad_mask)])
+    if labels is not None:
+        final_labels = torch.stack([torch.cat((pad_label, label)) for label, pad_label in zip(labels, pad_labels)])
+    return final_input_embeds, final_attention_mask, final_labels
 
 
 TRANSFORMERS_MODELS_TO_LORA_TARGET_MODULES_MAPPING = {