# What does this PR do?

This is a long pending followup to #1093 . This is meant to be a draft PR, and I'm hoping for comments from maintainers on the best way forward. 


To summarize, the current `PeftModelForCausalLM` code has different behaviour from `PeftModelForSeq2SeqLM` and the original prompt tuning implementation. The prompt embeddings are meant to be placed right before the input query, but with left padding, you end up with something like 
`Prompt<pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad><pad>Query`

The original prompt tuning method was proposed with T5, a seq2seq model, for which this issue isn't there. 

It is very messy to deal with this at the model stage! An ad-hoc fix is to detect the boundary between padding and non-padding tokens via the attention mask, and then split up the input batch of tensors into two. The prompt embeddings get added right before the input embeddings without padding. We can finally add the prompt tokens/embeddings back. I've had to break everything down into lists of tensors, because padding is going to be non-uniform. All the current tests pass for this implementation (more should be added if we decide to use this).

Another solution is to simply add a warning for the user for CausalLM + prompt tuning (this is, not great, because warnings usually get lost when your stack is PEFT + Accelerate + Transformers + PyTorch + NCCL). 

I am yet to do a proper benchmark for performance difference b/w this corrected version and current PEFT implementation. 
( When I ran the example notebook for GPT-2 on Twitter Complaints, there was almost no difference in performance, but then the Twitter Complaints dataset only has 50 train examples)


cc @BenjaminBossan @pacman100