@SumanthRH 2023-12-13
TODO:

- [ ] Bypass special handling with `attention_mask` is `None`.
- [ ] Add unit tests and asserts for `separate_pad_tokens` and `add_pad_tokens`.
- [ ] Performance comparison for current prompt tuning impl. and the corrected prompt tuning impl.