@SumanthRH 2023-12-23
Hi @BenjaminBossan , finally got some numbers! TLDR is again that performance differences on small datasets isn't much (and even the speed differences weren't much). 

I've benchmarked on subsets of the ELI5 and the OpenOrca dataset.

I used GPT2 on ELI5 (hoping this isn't in the training set). Eval numbers with the exact same hyperparameters on the current and the new, fixed implementation are below:

| Implementation    | train_ppl | train_epoch_loss | eval_ppl | eval_epoch_loss |
|--------------------|-----------|------------------|----------|-----------------|
| Old Implementation | 47.9587   | 3.8703           | 39.1261  | 3.6668          |
| New Implementation | 47.8425   | 3.8679           | 39.1270  | 3.6668          |

Not very significant imo. I thought maybe in this case the task was also not hard, so I benchmarked on a harder dataset: OpenOrca. 


I switched to the openlm-research/open_llama_3b model, as this is a harder task overall. Results after finetuning on  a 10K subset: 

| Implementation    | train_ppl | train_epoch_loss | eval_ppl | eval_epoch_loss |
|--------------------|-----------|------------------|----------|-----------------|
| Old Implementation | 3.4938    | 1.2510           | 3.4668   | 1.2432          |
| New Implementation | 3.4785    | 1.2466           | 3.4382   | 1.2349          |

Can see a second decimal point difference now with `eval_ppl` 🙆🏻‍♂️

I believe more benchmarking on larger datasets will make this more significant. From a research perspective, this improvement should still be useful

## Slowdown
I added time.time() statements for the input preprocessing in `PeftModelForCausalLM` . This is the time to make the new input_embeds, attention_mask and labels, right from [this start](https://github.com/huggingface/peft/blob/799420aef18146cffe98db7eeb3e4b34dc566110/src/peft/peft_model.py#L1083) to [this end](https://github.com/huggingface/peft/blob/799420aef18146cffe98db7eeb3e4b34dc566110/src/peft/peft_model.py#L1120) line number. I got the time deltas for the Orca dataset with a batch size of 32 (max sequence length of 512). Tests done on an RTX 2080 Ti. Taking the average over 40 iterations, I see:
Old implementation:  0.00061834 s ( 0.6 ms)
New implementation: 0.00955 s (9 ms)

There's a 15x slowdown in the preprocessing step, but it is still in the milliseconds even with a batch size of 32! From my perspective, this slowdown can be tolerated, but I'm curious what you think.