@pacman100 2024-02-19
> The bigger cost IMO is the cost of maintaining the more complex code. I'm really on the fence when it comes to this, as the code is significantly harder to maintain -- even if correct in theory, there appears to be no impact in practice.

I also feel the same.